Clean up op-chain-ops
