+++ op-chain-ops/Makefile
@@ -1,5 +1,7 @@
-op-migrate:
-	go build -o ./bin/op-migrate ./cmd/op-migrate/main.go
+all: check-l2
+
+check-l2:
+	go build -o ./bin/check-l2 ./cmd/check-l2/main.go
 
 test:
 	go test ./...
@@ -10,4 +12,4 @@ fuzz:
 	go test -run NOTAREALTEST -v -fuzztime 10s -fuzz=FuzzAliasing ./crossdomain
 	go test -run NOTAREALTEST -v -fuzztime 10s -fuzz=FuzzVersionedNonce ./crossdomain
 
-.PHONY: op-migrate test
+.PHONY: check-l2 test fuzz
